web_app/requirements.txt
pandas
streamlit

azure-core
azure-identity
azure-storage-blob
